+++ ankor-spring/src/main/java/at/irian/ankor/spring/SpringBasedBeanFactory.java
@@ -19,21 +19,25 @@ package at.irian.ankor.spring;
 import at.irian.ankor.annotation.AnnotationBeanMetadataProvider;
 import at.irian.ankor.ref.Ref;
 import at.irian.ankor.viewmodel.factory.ReflectionBeanFactory;
+import at.irian.ankor.viewmodel.metadata.BeanMetadata;
 import org.springframework.beans.BeansException;
-import org.springframework.beans.factory.config.AutowireCapableBeanFactory;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.ApplicationContextAware;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Scope;
+import org.springframework.stereotype.Component;
 
 /**
+ * Implementation of a {@link at.irian.ankor.viewmodel.factory.BeanFactory BeanFactory} providing basic Spring
+ * IoC support.
+ * This factory creates new bean instances exactly like the {@link at.irian.ankor.viewmodel.factory.ReflectionBeanFactory}.
+ * Right after the raw instance is created (and before it is initialized) Spring's auto-wiring is invoked, so that
+ * Spring beans may get injected into the newly created view model bean.
+ * After the Ankor initialization (i.e. view model post processing, see {@link at.irian.ankor.viewmodel.ViewModelPostProcessor})
+ * Spring's bean initialization is invoked, so that Spring specific
+ * {@link org.springframework.beans.factory.config.BeanPostProcessor BeanPostProcessors} are called.
+ *
  */
-@SuppressWarnings({"SpringFacetCodeInspection", "UnusedDeclaration"})
-@Configuration
+@Component
 public class SpringBasedBeanFactory extends ReflectionBeanFactory implements ApplicationContextAware {
-
-    private static final ThreadLocal<Ref> TEMPORARY_REF_HOLDER = new ThreadLocal<Ref>();
 
     private ApplicationContext applicationContext;
 
@@ -46,27 +50,21 @@ public class SpringBasedBeanFactory extends ReflectionBeanFactory implements App
         this.applicationContext = applicationContext;
     }
 
-    @Bean
-    @Scope("prototype")
-    public Ref ref() {
-        return TEMPORARY_REF_HOLDER.get();
-    }
-
-
     @SuppressWarnings("unchecked")
     @Override
-    protected <T> T createInstance(Class<T> type, Ref ref, Object[] constructorArgs) {
+    protected <T> T createRawInstance(Class<T> type, Ref ref, Object[] constructorArgs) {
+        T bean = super.createRawInstance(type, ref, constructorArgs);
+        applicationContext.getAutowireCapableBeanFactory().autowireBean(bean);
+        return bean;
 
-        AutowireCapableBeanFactory autowireCapableBeanFactory = applicationContext.getAutowireCapableBeanFactory();
+    }
 
-        T bean = super.createInstance(type, ref, constructorArgs);
-
-        TEMPORARY_REF_HOLDER.set(ref);
-        try {
-            autowireCapableBeanFactory.autowireBean(bean);
-            return (T)autowireCapableBeanFactory.initializeBean(bean, ref.propertyName());
-        } finally {
-            TEMPORARY_REF_HOLDER.remove();
-        }
+    @SuppressWarnings("UnnecessaryLocalVariable")
+    @Override
+    public Object initializeInstance(Object instance, Ref ref, BeanMetadata metadata) {
+        Object ankorInitializedBean = super.initializeInstance(instance, ref, metadata);
+        Object springInitializedBean = applicationContext.getAutowireCapableBeanFactory()
+                                                         .initializeBean(ankorInitializedBean, ref.propertyName());
+        return springInitializedBean;
     }
 }